Moved MangooUtils functions to Application class
