+++ mangooio-core/src/main/java/io/mangoo/core/Application.java
@@ -2,11 +2,17 @@ package io.mangoo.core;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
 import com.google.inject.Injector;
 
+import io.mangoo.cache.Cache;
+import io.mangoo.cache.GuavaCache;
+import io.mangoo.cache.HazlecastCache;
 import io.mangoo.configuration.Config;
+import io.mangoo.enums.Default;
 import io.mangoo.enums.Mode;
 
 /**
@@ -16,6 +22,7 @@ import io.mangoo.enums.Mode;
  *
  */
 public final class Application {
+    private static volatile Cache cache;
     private static volatile Config config;
     private static volatile Mode mode;
     private static volatile Injector injector;
@@ -141,5 +148,27 @@ public final class Application {
         Objects.requireNonNull(clazz, "clazz can not be null");
 
         return injector.getInstance(clazz);
+    }
+    
+    /**
+     * @return An instance of the internal mangoo I/O cache
+     */
+    public static Cache getInternalCache() {
+        if (cache == null) {
+            if (Default.CACHE_CLASS.toString().equals(config.getCacheClass())) {
+                cache = new GuavaCache();                
+            } else {
+                cache = new HazlecastCache();
+            }
+        }
+        
+        return cache;
+    }
+    
+    /**
+     * @return A list of all administrative URLs
+     */
+    public static List<String> getAdministrativeURLs() {
+        return Arrays.asList("@cache", "@metrics", "@config", "@routes", "@health", "@scheduler", "@memory", "@system");  
     }
 }
